Modified implementation of Composition.is_treble_fixed
Currently this returns True for single-part compositions as the treble comes
home at the end. This means we don't highlight the treble in rollups. This makes
sense if we ring more than one part as the treble won't rotate with the other
bells, but doesn't work when we're only ringing one part.

This commit fixes this situation. It still isn't quite right based on the name:
the treble's fixity surely ought to depend on whether the methods are treble-
dominated or otherwise, but it does do what we need for this project.